+++ src/client/kotlin/com/asc/AscKeybinding.kt
@@ -1,5 +1,6 @@
 package com.asc
 
+import com.asc.ui.AscScreen
 import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents
 import net.fabricmc.fabric.api.client.keybinding.v1.KeyBindingHelper
 import net.minecraft.client.MinecraftClient
@@ -8,9 +9,9 @@ import net.minecraft.client.util.InputUtil
 import org.lwjgl.glfw.GLFW
 
 object AscKeybinding {
-    private val TOGGLE_ITEM_FRAME_VISIBILITY = KeyBinding("key.asc.toggle_item_frame_visibility", InputUtil.Type.KEYSYM, GLFW.GLFW_KEY_UNKNOWN, "category.asc.armor_stand")
-    private val OPEN_GUI = KeyBinding("key.asc.open_gui", InputUtil.Type.KEYSYM, GLFW.GLFW_KEY_COMMA, "category.asc.armor_stand")
-    private val CHECK_ARMOR_STAND = KeyBinding("key.asc.check_armor_stand", InputUtil.Type.KEYSYM, GLFW.GLFW_KEY_UNKNOWN, "category.asc.armor_stand")
+    private val TOGGLE_ITEM_FRAME_VISIBILITY = KeyBinding("asc.key.toggle_item_frame_visibility", InputUtil.Type.KEYSYM, GLFW.GLFW_KEY_UNKNOWN, "asc.category.armor_stand")
+    private val OPEN_GUI = KeyBinding("asc.key.open_gui", InputUtil.Type.KEYSYM, GLFW.GLFW_KEY_COMMA, "asc.category.armor_stand")
+    private val CHECK_ARMOR_STAND = KeyBinding("asc.key.check_armor_stand", InputUtil.Type.KEYSYM, GLFW.GLFW_KEY_UNKNOWN, "asc.category.armor_stand")
 
     fun registerKeybindings() {
         KeyBindingHelper.registerKeyBinding(TOGGLE_ITEM_FRAME_VISIBILITY)
@@ -26,11 +27,11 @@ object AscKeybinding {
         }
 
         while (OPEN_GUI.wasPressed()) {
-            client.setScreen(AscScreen(client))
+            client.setScreen(AscScreen())
         }
 
         while (CHECK_ARMOR_STAND.wasPressed()) {
-            AscTriggers.check(client)
+            AscTriggers.check()
         }
     }
 }