+++ CHANGELOG.md
@@ -1,3 +1,5 @@
-### Armor Statues Companion v1.2.0+1.20.1
+### Armor Statues Companion v1.2.1+1.20.1
 
-* Updated to Minecraft 1.20.1
+* UI overall:
+  * Switch to left alignment
+  * Added colored buttons